Extract API endpoint modules
